requirements.txt
# Core FastAPI framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# AI Agents
pydantic-ai==0.0.9
openai==1.6.1
google-generativeai==0.3.2

# Voice Services
elevenlabs==0.2.27

# Environment and utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Development tools
httpx==0.25.2  # For testing
pytest==7.4.3
pytest-asyncio==0.21.1 